requirements.txt
fuzzywuzzy==0.17.0
matplotlib==3.1.1
opencv_contrib_python==3.4.2.17
pandas==0.25.0
torch==1.1.0
torchvision==0.3.0
six==1.12.0
tqdm==4.32.2
numpy==1.17.0
imutils==0.5.3
python_Levenshtein==0.12.0
Pillow==6.2.0
